ci(build): update cpan module cache path and key
The cache path for CPAN modules is updated from 'C:\strawberry\perl\site' to 'C:\strawberry\perl\vendor' to align with modern Perl standards.

The cache key is also updated to be more specific, including the versions of cpanm and PAR::Packer. This will prevent stale cache issues when these dependencies are updated. The redundant restore-keys are also removed.